Removing support for old legal hold APIs.
