+++ docs/methoddocs/legalhold.md
@@ -1,7 +1,5 @@
 # Legal Hold
 
-*For legal hold commands compatible with API Client authentication, see [Legal Hold - Api Clients](legalholdapiclient.md).*
-
 ```{eval-rst}
 .. autoclass:: pycpg.services.legalhold.LegalHoldService
     :members:
